06-app-server/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.3
pydantic-settings>=2.2.1
python-dotenv>=1.0.0
openai>=1.12.0 